+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased changes can be added here. -->
  - Updated sample file help text (EDPUB-780)
+ - Updated science value question (EDPUB-774)
  - Updated data format question text (EDPUB-778)
  - Updated Node version to v18.14.1